usb: musb: ti81xx: auto vbus control across host and device mode
Automatically control vbus using timer. Set the session bit periodically in a timer
context when nothing is connected to usb port. If mini-A is connected then Vbus will
get switched on and if mini-B is connected then Vbus will nto be driven.

Kill the timer when a device is attached in host mode or port is connected to external
host and revive the timer as soon as disconnect happens.

Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>