davinci: vpif capture: declare contiguous region of memory handled by dma_alloc_coherent
add support to declare contiguous region of memory to be handled
when requested by dma_alloc_coherent when it's asked for coherent
memory for this device. The user can specify the size of the buffers
with an offset from the kernel image using cont_bufsize and
cont_bufoffset module parameters respectively.

Signed-off-by: Manjunath Hadli <manjunath.hadli@ti.com>