da830/omap-l137: Configure LCD_AC_ENB_CS pin as LCD for Char LCD
configured the LCD_AC_ENB_CS pin as GPIO for Graphical
LCD to work on AM1707 parts. But this had caused Character
LCD not to work which requires the LCD_AC_ENB_CS pin to be
functioning as LCD. This patch configures the LCD_AC_ENB_CS
pin as LCD incase Character LCD is enabled.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>